ckb-auth-types merge to ckb-auth-rs
