+++ ckb-auth-rs/Cargo.toml
@@ -5,8 +5,7 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[dependencies]
-ckb-auth-types = { path = "../ckb-auth-types" }
+[target.'cfg(target_arch = "riscv64")'.dependencies]
 ckb-std = { version = "0.14.3", features = ["ckb2023"] }
-log = { version = "0.4.17", default-features = false }
+log = { version = "0.4.17", default-features = false}
 hex = { version = "0.4.3", default-features = false, features = ["alloc"]}
