+++ tests/auth-c-tests/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-ckb-auth-types = { path = "../../ckb-auth-types" }
+ckb-auth-rs = { path = "../../ckb-auth-rs" }
 ckb-crypto = "0.111.0"
 ckb-error = "0.111.0"
 ckb-hash = "0.111.0"
