+++ tests/auth-c-tests/src/tests/mod.rs
@@ -1,7 +1,7 @@
 #![allow(unused_imports)]
 #![allow(dead_code)]
 
-use ckb_auth_types::EntryCategoryType;
+use ckb_auth_rs::EntryCategoryType;
 use ckb_chain_spec::consensus::{Consensus, ConsensusBuilder};
 use ckb_crypto::secp::{Generator, Privkey, Pubkey};
 use ckb_types::{
