+++ tools/ckb-auth-cli/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-ckb-auth-types = { path = "../../ckb-auth-types" }
+ckb-auth-rs = { path = "../../ckb-auth-rs" }
 ckb-vm = { version = "0.24.0", features = [ "asm" ] }
 ckb-types = "=0.111.0-rc2"
 ckb-hash = "=0.111.0-rc2"
