+++ tools/ckb-auth-cli/src/chain_command/bitcoin.rs
@@ -1,6 +1,6 @@
 use crate::{auth_script::run_auth_exec, utils::decode_string, BlockChain, BlockChainArgs};
 use anyhow::{anyhow, Error};
-use ckb_auth_types::AuthAlgorithmIdType;
+use ckb_auth_rs::AuthAlgorithmIdType;
 use clap::{arg, ArgMatches, Command};
 
 pub struct BitcoinLockArgs {}
