+++ tools/ckb-auth-cli/src/chain_command/ripple.rs
@@ -4,7 +4,7 @@ use crate::{
     BlockChain, BlockChainArgs,
 };
 use anyhow::{anyhow, Error};
-use ckb_auth_types::AuthAlgorithmIdType;
+use ckb_auth_rs::AuthAlgorithmIdType;
 use clap::{arg, ArgMatches, Command};
 use hex::decode;
